chore: 폴더 구조 변경 (hooks -> queries, libs -> services)
