+++ components/common/header/MobileList.tsx
@@ -2,13 +2,13 @@ import { useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
 import Link from 'next/link';
 import { useUserStore } from '@/store/authStore';
-import { useUser } from '@/hooks/useUser';
 import useModalStore from '@/store/useModalStore';
 import TeamForm from '../modal/TeamForm';
+import { useUsersQuery } from '@/queries/user/user';
 
 export default function MobileList() {
   const { accessToken } = useUserStore();
-  const { data: user } = useUser(accessToken);
+  const { data: user } = useUsersQuery(accessToken);
   const openModal = useModalStore((state) => state.openModal);
 
   const teams = user?.memberships.map((membership) => membership.group) || [];
