+++ components/team/member/MemberList.tsx
@@ -10,8 +10,8 @@ import useModalStore from '@/store/useModalStore';
 import CopyEmail from '@/components/common/modal/CopyEmail';
 import ConfirmModal from '@/components/common/modal/ConfirmModal';
 import { toast } from 'react-toastify';
-import { useDeleteMember } from '@/hooks/useDeleteMember';
 import { useUserStore } from '@/store/authStore';
+import { useDeleteMemberQuery } from '@/queries/group/member';
 
 interface MemberProps {
   member: Member;
@@ -23,7 +23,7 @@ export default function MemberList({ member, isAdmin }: MemberProps) {
   const menuRef = useRef<HTMLDivElement>(null);
   const openModal = useModalStore((state) => state.openModal);
 
-  const deleteMemberMutation = useDeleteMember(member.groupId);
+  const deleteMemberMutation = useDeleteMemberQuery(member.groupId);
 
   const { user } = useUserStore();
   const currentUserId = user?.id;
@@ -89,10 +89,16 @@ export default function MemberList({ member, isAdmin }: MemberProps) {
           className="rounded-full object-contain"
         />
         <div className="flex min-w-0 flex-col">
-          <span className="font-medium-14 flex flex-row gap-x-1 text-text-primary">
+          <span className="font-medium-14 flex flex-row gap-x-1 truncate text-text-primary">
             {member.userName}
             {member.role === 'ADMIN' && (
-              <Image src={crownIcon} alt="관리자" width={14} height={17} />
+              <Image
+                src={crownIcon}
+                alt="관리자"
+                width={14}
+                height={17}
+                className="object-contain"
+              />
             )}
           </span>
           <span className="font-regular-12 truncate text-text-secondary">
